+++ .github/workflows/build.yml
@@ -37,7 +37,13 @@ jobs:
         run: cargo fmt --all -- --check
       - name: Build
         run: cargo check --all
+      - name: Clippy
+      - uses: giraffate/clippy-action@v1
+        with:
+         reporter: 'github-pr-review'
+         github_token: ${{ secrets.GITHUB_TOKEN }}       
       - name: Test
         env:
           TEST_PRIVATE_KEY: ${{ secrets.TEST_PRIVATE_KEY }}
         run: cargo test --all
+
